+++ src/app/common/components/login/login.component.ts
@@ -1,4 +1,10 @@
-import { Component, OnInit, ChangeDetectionStrategy } from '@angular/core';
+import { Component, OnInit, OnDestroy, ChangeDetectionStrategy } from '@angular/core';
+
+import { Router } from '@angular/router';
+
+import { Subscription } from 'rxjs/Subscription';
+
+import { AuthService } from '../../../core';
 
 @Component({
   selector: 'app-login',
@@ -6,11 +12,45 @@ import { Component, OnInit, ChangeDetectionStrategy } from '@angular/core';
   styleUrls: ['./login.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class LoginComponent implements OnInit {
+export class LoginComponent implements OnInit, OnDestroy {
+  message: string;
 
-  constructor() { }
+  private sub: Subscription;
+
+  constructor(
+    public authService: AuthService,
+    public router: Router
+  ) {}
 
   ngOnInit() {
+    this.setMessage();
   }
 
+  ngOnDestroy() {
+    if (this.sub) {
+      this.sub.unsubscribe();
+    }
+  }
+
+  login() {
+    this.message = 'Trying to log in ...';
+    this.sub = this.authService.login().subscribe(() => {
+      this.setMessage();
+      if (this.authService.isLoggedIn) {
+        const redirect = this.authService.redirectUrl
+          ? this.authService.redirectUrl : '/admin';
+
+        this.router.navigate([redirect]);
+      }
+    });
+  }
+
+  logout() {
+    this.authService.logout();
+    this.setMessage();
+  }
+
+  private setMessage() {
+    this.message = 'Logged ' + (this.authService.isLoggedIn ? 'in as admin' : 'out');
+  }
 }