+++ src/app/core/guards/auth.guard.ts
@@ -1,0 +1,39 @@
+import { Injectable } from '@angular/core';
+import { CanActivate, CanLoad, ActivatedRouteSnapshot, RouterStateSnapshot, Route, Router } from '@angular/router';
+
+import { Observable } from 'rxjs/Observable';
+
+import { AuthService } from '../services/auth.service';
+
+@Injectable()
+export class AuthGuard implements CanActivate, CanLoad {
+
+  constructor(
+    private authService: AuthService,
+    private router: Router
+  ) { }
+
+  canActivate(
+    next: ActivatedRouteSnapshot,
+    state: RouterStateSnapshot): Observable<boolean> | Promise<boolean> | boolean {
+    return true;
+  }
+
+  canLoad(route: Route): Observable<boolean> | Promise<boolean> | boolean {
+    const url = `/${route.path}`;
+
+    return this.checkLogin(url);
+  }
+
+  private checkLogin(url: string): boolean {
+    if (this.authService.isLoggedIn) {
+      return true;
+    }
+
+    this.authService.redirectUrl = url;
+
+    this.router.navigate(['/login']);
+
+    return false;
+  }
+}