+++ src/app/core/index.ts
@@ -1,0 +1,9 @@
+export * from './services/auth.service';
+export * from './services/cart.service';
+export * from './services/config-options.service';
+export * from './services/constants.service';
+export * from './services/generator.service';
+export * from './services/local-storage.service';
+export * from './services/generete-n-sequence.factory';
+
+export * from './guards/auth.guard';