+++ src/app/core/services/index.ts
@@ -1,5 +1,0 @@
-export * from './config-options.service';
-export * from './constants.service';
-export * from './generator.service';
-export * from './local-storage.service';
-export * from './generete-n-sequence.factory';