+++ src/app/products/components/product-list/product-list.component.ts
@@ -5,7 +5,7 @@ import { Observable } from 'rxjs/Observable';
 
 import { IProduct } from '../../../shared/interfaces';
 import { ProductService } from '../../products.service';
-import { CartService } from '../../../core/cart.service';
+import { CartService } from '../../../core/';
 
 @Component({
   selector: 'app-product-list',